Fitur download member budrest

See merge request my-archery/api!1624

+++ database/migrations/2022_08_12_145236_alter_table_venue_products.php
@@ -13,13 +13,17 @@ class AlterTableVenueProducts extends Migration
      */
     public function up()
     {
-        Schema::rename('venue_place_product_customs', 'venue_place_products');
+        if (Schema::hasTable('venue_place_product_customs')) {
+            Schema::rename('venue_place_product_customs', 'venue_place_products');
+        }
 
-        Schema::table('venue_place_products', function (Blueprint $table) {
-            $table->string("description")->nullable()->after("product_name");
-            $table->dropColumn('total_quota_per_day');
-            $table->boolean("has_session")->default(false)->after("weekend_price");
-        });
+        if (Schema::hasTable('venue_place_products')) {
+            Schema::table('venue_place_products', function (Blueprint $table) {
+                $table->string("description")->nullable()->after("product_name");
+                $table->dropColumn('total_quota_per_day');
+                $table->boolean("has_session")->default(false)->after("weekend_price");
+            });
+        }
     }
 
     /**
